pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "url-vault"
version = "0.1.0"
description = "Build a local cache of mirrored Git repositories and URLs."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.3",
    "rich-click>=1.9.7",
]

[project.scripts]
url-vault = "url_vault.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.setuptools]
packages = ["url_vault"]

[tool.pytest.ini_options]
addopts = "--basetemp=.pytest-tmp -p no:cacheprovider"
testpaths = [
    "tests",
]
